% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeter.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check for required packages based on selected methods}
\usage{
check_dependencies(binning_method, woe_post_cluster)

check_dependencies(binning_method, woe_post_cluster)
}
\arguments{
\item{binning_method}{character - binning method selected}

\item{woe_post_cluster}{logical - whether post-clustering is used}
}
\description{
Check for required packages based on selected methods

Check for required packages based on selected methods
}
