% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_naming_conventions.R
\name{check_naming_conventions}
\alias{check_naming_conventions}
\title{Check on the naming convention}
\usage{
check_naming_conventions(data, test_target_type = FALSE)
}
\arguments{
\item{data}{data.table or data.frame}
}
\value{
a table which contains two columns "ErrorType" and "ErrorTarget". Each row corresponds to a variable of the data.
The column "ErrorType" == FALSE when the condition on the first letter is respected.
The column "ErrorTarget" == FALSE when the condition on BIN/INT/ORD/CAT on the target is respected
}
\description{
This function determines if the naming convention is respected for each column.
}
