% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeter.r
\name{compute_target_statistics}
\alias{compute_target_statistics}
\title{Compute statistics for target variable}
\usage{
compute_target_statistics(data, target, target_type)

compute_target_statistics(data, target, target_type)
}
\arguments{
\item{data}{data.table with the data}

\item{target}{character - name of the target variable}

\item{target_type}{character - type of the target variable}
}
\value{
data.table with target statistics

data.table with target statistics
}
\description{
This function computes appropriate statistics based on the target type

This function computes appropriate statistics based on the target type
}
