% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeter.r
\name{create_binning_expression}
\alias{create_binning_expression}
\title{Create expression for binning data}
\usage{
create_binning_expression(target, num_vars, other_vars, binning_function)
}
\arguments{
\item{target}{character - name of the target variable}

\item{num_vars}{character vector - names of numeric variables}

\item{other_vars}{character vector - names of other variables}

\item{binning_function}{function - the binning function to use}
}
\value{
character string with the expression to evaluate
}
\description{
Helper function to create the expression for binning numeric variables
}
