% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvar.R
\name{crossvar}
\alias{crossvar}
\title{Cross a single variable with a target}
\usage{
crossvar(data, target, var, ...)
}
\arguments{
\item{data}{data.frame or data.table}

\item{target}{character - name of the target variable of interest}

\item{var}{character - name of the candidate variable to be crossed with target}

\item{...}{Additional parameter to be passeed to \code{\link[targeter]{targeter}}}
}
\value{
returns an object of class crossvar
}
\description{
this is a convenience function that behind the scene calls complete \code{\link[targeter]{targeter}}
 function on data. This allows to quickly assess relation for one variable and have access to plots. It is highly
 recommended to have a look at \code{\link[targeter]{targeter}} parameters as they could be used here also.
}
\details{
content of returned object depends on the type of the target (binary/categorical/continuous.)
 For more information, look at the documentation of the crosssvar class.
}
\examples{
\dontrun{
if(interactive()){
 data(adult)
 crossvar(adult, target='ABOVE50K', var='AGE')
 t <- crossvar(adult, target='HOURSPERWEEK', var='FNLWGT')
 plot(t)
 }
}
}
\seealso{
\code{\link[targeter]{targeter}}
 \code{\link[plot.crossvar]{plot.crossvar}}
}
