% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{plot.tartree}
\alias{plot.tartree}
\title{Plot a targeter tartree object}
\usage{
\method{plot}{tartree}(
  x,
  format = "static",
  prefix = "target = ",
  type = 2,
  yesno = 2,
  branch = 0,
  branch.type = 5,
  box.palette = "Blues",
  shadow.col = 0,
  cex = 0.7,
  left = FALSE,
  ...
)
}
\arguments{
\item{x}{- a decision tree as fitted by `tartree` function}

\item{format}{see `rpart.plot` help}

\item{prefix}{-see `rpart.plot` help}

\item{type}{One of 'static' (use rpart.plot) or interactive (use visTree)
Parameter expansion is used so that one can use "i". Default: static.}

\item{yesno}{-see `rpart.plot` help}

\item{branch}{-see `rpart.plot` help}

\item{branch.type}{-see `rpart.plot` help}

\item{box.palette}{-see `rpart.plot` help}

\item{shadow.col}{-see `rpart.plot` help}

\item{cex}{-see `rpart.plot` help}

\item{left}{-see `rpart.plot` help}

\item{...}{-  eventuial additional parameters to be passed to `rpart.plot`
or visTree}
}
\description{
This function creates a plot for a targeter decision tree 
model build with `tartree` function If format is static, one will use
rpart.plot package. If interactive, one will use visTree.
}
