% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_woe}
\alias{plot_woe}
\title{WOE graph for binary/continuous targets}
\usage{
plot_woe(
  x,
  metadata = NULL,
  print_NA = TRUE,
  numvar_as = c("bin", "value"),
  title = TRUE,
  do_plot = TRUE,
  ...
)
}
\arguments{
\item{x}{a crossvar object, as found in profiles slot of a targeter object.}

\item{metadata}{data.frame - if metadata is  loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available.
The label will be used for the title and the x-axis of the graph.}

\item{print_NA}{boolean. Should we display WOE for NA values of the explanatory variable. Default: TRUE}

\item{numvar_as}{character. How should continuous explanatory variables be displayed:
#'\itemize{
\item bin (default) - display WOE for adjacent binning without respecting variable raw values
\item value - plot WOE using centers of binning classes, thus respecting variable raw values
}}

\item{title}{boolean: do we plot default generated title Y/N or character (override title) - default: TRUE}

\item{do_plot}{boolean - whether to effectively show the plot or not (internal use to combine plots)}

\item{...}{extra parameters (not used currently)}
}
\value{
plot a graph and returns its ggplot2 object
}
\description{
For binary/continuous targets, WOE (Weight of Evidence) statistics are computed.
This function plots a WOE barplot (or lines) for an explanatory variable, based on a crossvar object.
}
\examples{
\dontrun{
if(interactive()){
 t <- crossvar(adult, target='ABOVE50K', var='AGE')
 plot_woe(t)
 plot_woe(t, numvar_as='value')
 }
}
}
\seealso{
\code{\link[targeter]{crossvar}}
}
