% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.targeter}
\alias{print.targeter}
\title{print.targeter}
\usage{
\method{print}{targeter}(x, ...)
}
\arguments{
\item{x}{targeter object}

\item{...}{additional arguments}
}
\description{
Print method for targeter objects
}
\details{
This method prints the targeter object, including the target, data name, date, and number of profiles available.
}
