% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{report}
\alias{report}
\title{report}
\usage{
report(
  object,
  summary_object = NULL,
  metadata = NULL,
  format = "html",
  nmax = 100,
  template = NULL,
  pptx_reference_doc = NULL,
  output_dir = paste0("targeter-report", format(Sys.time(), format =
    "\%Y-\%m-\%d_\%H\%M\%S")),
  output_file = "index",
  title = object$analysis,
  author = getOption("targeter.author", "welovedatascience targeter"),
  fullplot_which_plot = "1:2",
  tables = FALSE,
  brandfile = getOption("targeter.brandfile", file.path(find.package("targeter", lib.loc
    = .libPaths()), "ressources", "_brand.yml")),
  logo = NULL,
  debug = FALSE,
  render = TRUE,
  ...
)
}
\arguments{
\item{object}{- an object of class targeter}

\item{summary_object}{an object of class "summary.targeter": pre-computed
summary of object.}

\item{metadata}{- data.frame - if metadata is  loaded in R environment,
label of the variables can be used. Default value (NULL) corresponds to
no metadata available.
The label will be used for the title and the x-axis of the graph.}

\item{format}{- Target output format (defaults to "html").
The option "all" will render all formats.}

\item{nmax}{- integer: maximum number of variables to be used. Default 100. 
use NULL to force all variables to be present in report.See notes.}

\item{template}{- character. path to a Quarto qmd document that will be used
to generate presentation. If NULL (default), we will use targeter package
default template.}

\item{pptx_reference_doc}{- character. Path to a default PPTX template 
that will be used as Quarto reference-doc for pppx formzt
If NULL (default), we will use targeter
package default templates for PPTX. 
If empty string "", we won't use any template.}

\item{output_dir}{- character. Output directory - default to create a 
new dedicated directory in working folder.}

\item{output_file}{- character, name of output file (without file extension).
default value: index}

\item{title}{- character, title for the presentation. Overide default title
that consists in taking title for targeter object 'analysis' slot.}

\item{author}{- character: presentatiopn author.}

\item{fullplot_which_plot}{- character, selection of plots (cf `fullplot`
documentation)}

\item{tables}{- logical whetheer to print tables in seprate slides in 
addition of graphics. Might ot produce nice outputs as tables are usually
large.}

\item{logo}{- character: path to a logo file. If NULL (default), we will take
package welovedatascience logo. If empty, no logo will be used. Only used
by revealjs format.}

\item{debug}{- logical. If TRUE, the function will print debug information (passed to Quarto call).}

\item{render}{- logical. If TRUE (default), the report will be rendered
and saved in the output_dir. If FALSE, the report will be generated but not
rendered. Useful for debugging.}

\item{...}{additional parameters to be passed to quarto, for instance
allowing to pass/overwite any YAML set-up}
}
\value{
invisibly returns path to the generated specific file (unique format)
or folder (several formats)
}
\description{
This function creates an automatic report according to a
 predefined template in the package or a user-generated template.
Create a report from a targeter report object
}
\examples{
\dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
slidify(tar, format = "pptx", author = "this is me")
}
}
\seealso{
\itemize{
\item \code{\link{targeter}}
\item \code{\link{summary.targeter}}
\item \code{\link{focus}}
}
}
