% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Create a presentation from a targeter report object}
\usage{
slidify(
  object,
  metadata = NULL,
  format = "pptx",
  template = NULL,
  pptx_template = NULL,
  output_dir = ".",
  output_file = NULL,
  tmpdir = tempdir(),
  delete_temporary_files = TRUE,
  title = object$analysis,
  author = getOption("targeter.author", "welovedatascience targeter"),
  fullplot_which_plot = "1:2",
  tables = FALSE,
  logo = NULL,
  ...
)
}
\arguments{
\item{object}{- an object of class targeter}

\item{metadata}{- not used currently}

\item{format}{- character, one of "pptx", "revealjs"}

\item{template}{- character. path to a Quarto qmd document that will be used
to generate presentation. If NULL (default), we will use targeter package
default template.}

\item{output_dir}{- character. Output directory - default to working
directory}

\item{output_file}{- character, name of output file (with extension). if NULL
(default), slidify will generate a temporary name.}

\item{tmpdir}{- path to a temporarry folder tp handle quarto rendering, 
By default, a call to `tempdir()` is performed.}

\item{delete_temporary_files}{- logical. Whether we delete temporary files or
not once presentation  is generated. Default: TRUE. On might want to use
FALSE for debugging purpose.}

\item{title}{- character, title for the presentation. Overide default title
that consists in taking title for targeter object 'analysis' slot.}

\item{author}{- character: presentatiopn author.}

\item{fullplot_which_plot}{- character, selection of plots (cf `fullplot`
documentation)}

\item{tables}{- logical whetheer to print tables in seprate slides in 
addition of graphics. Might ot produce nice outputs as tables are usually
large.}

\item{logo}{- character: path to a logo file. If NULL (default), we will take
package welovedatascience logo. If empty, no logo will be used. Only used
by revealjs format.}

\item{...}{- not yet used.}

\item{reference_doc}{- character. Path to a default PPTX/reveal HTML template 
that will be used as Quarto reference-doc
(mostly for PPTX) If NULL (default), we will use targeter
package default templates for PPTX and no template for revealjs
If empty, default template will be used.}
}
\value{
inviisibly returns path to the generated presentation
}
\description{
This function generates a presentation (slides), currently targeting
powerpoint (other formats to come).
}
\examples{
\dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
slidfy(tar)
}
}
