% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_targeter.R
\name{summary.targeter}
\alias{summary.targeter}
\title{summary.targeter}
\usage{
\method{summary}{targeter}(
  object,
  extra_stats = FALSE,
  criteria = c("IV", "index.max.index", "chisquare", "pvalue", "index.max.count",
    "index.max.props"),
  ...
)
}
\arguments{
\item{object}{an object of class targeter}

\item{extra_stats}{boolean - whether function returns some additional associations statistics (default: FALSE)}

\item{criteria}{which criteria is used to sort the summary table (default: IV)}

\item{...}{additional parameters (currently not used)}
}
\value{
a data.table summary table
}
\description{
summary method for targeter object
}
\details{
summary method invoked on a targeter object loops over all individual profiles/crossvar/associations and derive statistics (cf \code{\link{summary.crossvar}}).
Returned table contains thus one row per explanatory variable analyzed in targeter and columns for associations measures / metrics.
}
\examples{
data(adult)
t <- targeter(adult,target ="ABOVE50K",analysis_name="Analyse")
summary(t)
summary(t, extra_stats = TRUE)
}
\seealso{
\code{\link{summary.crossvar}}
}
