% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{table.crossvar}
\alias{table.crossvar}
\title{table.crossvar: build a meaningful table from a crossvar object.}
\usage{
table.crossvar(x, round_digits = 3)
}
\arguments{
\item{x}{crossvar - crossvar object (stored in targeter$profiles slot).}

\item{round_digits}{integer - number of digits to rounds some values.}
}
\value{
a data.frame
}
\description{
This functions builds a dataframe that gather most relevant
information from a crossvar object. Useful in reports, associated with
kableExtra or other tables output generating engine.
}
\examples{
\dontrun{
tar <- targeter(adult, target = 'ABOVE50K')
table.crossvar(tar$profiles[['EDUCATION']])
}

}
