% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{table_crossvar}
\alias{table_crossvar}
\title{table_crossvar: build a meaningful table from a crossvar object.}
\usage{
table_crossvar(x, round_digits = 3, counts_and = c("props", "target\%"))
}
\arguments{
\item{x}{crossvar - crossvar object (stored in targeter$profiles slot).}

\item{round_digits}{integer - number of digits to rounds some values.}

\item{counts_and}{- character - output will always contain counts, we can 
add row percentages and column percentage for target (both always by default)}
}
\value{
a data.frame
}
\description{
This functions builds a dataframe that gather most relevant
information from a crossvar object. Useful in reports, associated with
kableExtra or other tables output generating engine.
}
\examples{
tar <- targeter(adult, target = 'ABOVE50K')
table_crossvar(tar$profiles[['EDUCATION']])

}
