% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeter.R
\name{targeter_big}
\alias{targeter_big}
\title{Process targeter by blocks for datasets with many variables}
\usage{
targeter_big(data, target, by_nvars = 250, verbose = TRUE, ...)
}
\arguments{
\item{data}{data - data.table or data.frame.}

\item{target}{character - name of the variable to explain.}

\item{by_nvars}{integer - number of variables to process in each group. Default is 250.}

\item{verbose}{boolean - if TRUE, prints progress information. Default is TRUE.}

\item{...}{additional parameters passed to the `targeter` function.}
}
\value{
A combined targeter object from all processed groups.
}
\description{
This function processes large datasets by splitting them into smaller groups and applying the `targeter` function to each group.
}
\examples{
\dontrun{
targeter_big(data, target = "ABOVE50K", by_nvars = 100)
}
}
