% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeter.r
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs for targeter function}
\usage{
validate_inputs(
  data,
  target,
  select_vars,
  exclude_vars,
  description_data,
  description_target,
  analysis_name,
  useNA,
  order_label,
  nbins
)
}
\arguments{
\item{data}{data.frame or data.table - data to analyze}

\item{target}{character - name of target variable}

\item{select_vars}{character vector - variables to select}

\item{exclude_vars}{character vector - variables to exclude}

\item{description_data}{character - data description}

\item{description_target}{character - target description}

\item{analysis_name}{character - name of analysis}

\item{useNA}{character - how to handle NA values}

\item{order_label}{character - method for ordering labels}

\item{nbins}{numeric - number of bins}
}
\value{
invisible(TRUE) if all validations pass
}
\description{
Validate inputs for targeter function
}
