% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{FUNCTION_TITLE}
\usage{
explore(object, summary_object = NULL, metadata = NULL, ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{summary_object}{PARAM_DESCRIPTION, Default: NULL}

\item{metadata}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[miniUI]{miniPage}}, \code{\link[miniUI]{miniTitleBar}}, \code{\link[miniUI]{miniButtonBlock}}, \code{\link[miniUI]{miniContentPanel}}
 \code{\link[shiny]{fluidPage}}, \code{\link[shiny]{column}}, \code{\link[shiny]{plotOutput}}, \code{\link[shiny]{renderPlot}}, \code{\link[shiny]{observeEvent}}, \code{\link[shiny]{runGadget}}, \code{\link[shiny]{viewer}}
 \code{\link[DT]{dataTableOutput}}
 \code{\link[gridExtra]{arrangeGrob}}
 \code{\link[targeter]{report}}
}
