% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{label}
\alias{label}
\title{Add label to a var}
\usage{
label(var, metadata, var_field = "var", label_field = "LABEL", lang = NULL)
}
\arguments{
\item{var}{character - the name of the var(s) that we want to remplace by the label, could be a vector}

\item{metadata}{data.frame - if we have a metadata loaded in your R, you can use to take the label of the vars.}

\item{lang}{character - the code for the language. It's useful if you have a metadata with different labels for differents languages.
By default, the value is NULL.}
}
\value{
a label
}
\description{
This functions allows to add a label to var. It's useful for the graphics or the reports.
Obviously, this function can be applied only if a metadata exists
}
\examples{
\dontrun{
label(adult$EDUCATIONNUM, metadata=mymetadata) # assumes mymetadata file exists
}
}
