% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.crossvar}
\alias{plot.crossvar}
\title{Function plot.crossvar()}
\usage{
\method{plot}{crossvar}(x, ...)
}
\arguments{
\item{x}{object of class "crossvar"}

\item{...}{other parameters}

\item{show}{character - determine on which data considered. By default, the value is "counts".
 This parameter can only take the following values:
\itemize{
\item "counts" : the data used is the contigency table created by the function crossvar.
\item "props" : the data used is the proportion table created by the function crossvar.
\item "index" : the data used is the index table created by the function crossvar.
}}

\item{type}{character - determine the type of graph. By default, the value is "auto".
 This parameter can only take the following values:
\itemize{
\item "auto" : "counts" shows a bar plot and "props" shows a line plot.
\item "bars" : shows a bar plot.
\item "line": shows a line plot.
}}

\item{metadata}{data.frame - if metadata is  loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available.
The label will be used for the title and the x-axis of the graph.}

\item{print_NA}{boolean - By default, the value is TRUE. If FALSE, the missing values of the variable are not printed.}

\item{target_NA}{boolean - By default, the value is TRUE. If FALSE, the missing values of the target are not printed.}

\item{only_target_ref_level}{boolean - By default, the value is FALSE. If TRUE, only print the values for the target TRUE.}

\item{lim_y}{boolean - By default, the value is TRUE. The axis y for the proportion is limited between 0 and 100.}
}
\value{
The function returns a graph.
}
\description{
This function allows to generate graphs on the object of class "crossvar".
}
\examples{
t <- crossvar(adult,target="ABOVE50K", var="AGE",)
plot(t,"counts")
plot(t,"counts",type="line")
plot(t,"props",print_NA = FALSE, only_target_ref_level = TRUE)
}
\seealso{
\itemize{
\item \code{\link{crossvar}}
\item \code{\link{summary.crossvar}}
\item \code{\link{cadran_plot.crossvar}}
}
}
