% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{quadrant_plot}
\alias{quadrant_plot}
\title{Function quadrant_plot}
\usage{
quadrant_plot(x, metadata = NULL, max_ncat = 15, print_NA = TRUE)
}
\arguments{
\item{x}{object of class "crossvar"}

\item{metadata}{data.frame - if metadata is  loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available.
The label will be used for the title and the x-axis of the graph.}
}
\value{
The function returns a graph.
}
\description{
This function allows to generate a quandrant graphic on an object of class "crossvar".
}
\examples{
t <- crossvar(adult,"ABOVE50K","WORKCLASS")
quadrant_plot(t)
}
\seealso{
\itemize{
\item \code{\link{crossvar}}
\item \code{\link{summary.crossvar}}
\item \code{\link{plot.crossvar}}
}
}
