% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(
  object,
  template = NULL,
  summary_object = NULL,
  browse = FALSE,
  ntop = NULL,
  nmin = 20,
  criteria = c("IV"),
  min_criteria = NULL,
  metadata = NULL,
  force_vars = character(),
  output_format = c("html", "pdf", "word"),
  output_file = NULL,
  output_dir = tempdir(),
  ...
)
}
\arguments{
\item{object}{an object of calss "targeter".}

\item{template}{template file path. By default, it put to NULL and take the default template.}

\item{summary_object}{an object of class "summary.targeter": pre-computed summary of object.}

\item{browse}{by default TRUE : print in the browse the output.}

\item{ntop}{integer - number of variables to show in the report. For more information, see the function focus By default, NULL i.e. all variables are taken.}

\item{nmin}{integer - minimum number of profiles in the seleected class. For more information, see the function focus.}

\item{criteria}{character - By default, the value is index.max.index. For more information, see the function focus.}

\item{min_criteria}{- By default NULL. If it's speciefied, only the observations which have criteria >= min_criteria.}

\item{metadata}{data.frame - if metadata is  loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available.
The label will be used for the title and the x-axis of the graph.}

\item{force_vars}{- character, list of variables that will be kept and then won't never be filterered.}

\item{output_format}{- By default it's html. For a pdf, output_format="pdf_document" and for a word, output_format="word_document".}

\item{output_file}{- name of the output file to be generated. If NULL (default) an, automated name will be generated.}

\item{output_dir}{- output directory, default: tempdir().}

\item{...}{other parameter from the function render.}
}
\value{
path to the generated report. Side effect would be to open a document, if \code{browse} is TRUE.
}
\description{
This function creates an automatic report according to a predefined template in the package or a user-generated template.
}
\examples{
\dontrun{
t <- targeter(adult,target ="ABOVE50K",analysis_name="Analyse",naming_conventions=FALSE)
report(t,output_format="pdf_document")
report(ntop=5)
}
}
\seealso{
\itemize{
\item \code{\link{targeter}}
\item \code{\link{summary.targeter}}
\item \code{\link{focus}}
}
}
