% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Create a presentation from a targeter report object}
\usage{
slidify(
  object,
  metadata = NULL,
  format = "pptx",
  template = NULL,
  pptx_reference_doc = NULL,
  revealjs_template = "",
  output_dir = ".",
  output_file = NULL,
  delete_temporary_files = TRUE,
  title = object$analysis,
  author = getOption("targeter.author", "welovedatascience targeter"),
  fullplot_which_plot = "1:2",
  tables = FALSE,
  logo = NULL,
  ...
)
}
\arguments{
\item{object}{- an object of class targeter}

\item{metadata}{- not used currently}

\item{format}{- character, one of "pptx", "revealjs"}

\item{template}{- character. path to a Quarto qmd document that will be used
to generate presentation. If NULL (default), we will use targeter package
default template.}

\item{pptx_reference_doc}{- character. Path to a default PPTX template 
that will be used as Q"uarto reference-doc
If NULL (default), we will use targeter
package default templates for PPTX. 
If empty string "", we won't use any template.}

\item{revealjs_template}{- character. Path to a default revealjs
template If empty string (default), no revealjs template will be used.}

\item{output_dir}{- character. Output directory - default to working
directory}

\item{output_file}{- character, name of output file (with extension). if NULL
(default), slidify will generate a temporary name.}

\item{delete_temporary_files}{- logical. Whether we delete temporary files or
not once presentation  is generated. Default: TRUE. On might want to use
FALSE for debugging purpose.}

\item{title}{- character, title for the presentation. Overide default title
that consists in taking title for targeter object 'analysis' slot.}

\item{author}{- character: presentatiopn author.}

\item{fullplot_which_plot}{- character, selection of plots (cf `fullplot`
documentation)}

\item{tables}{- logical whetheer to print tables in seprate slides in 
addition of graphics. Might ot produce nice outputs as tables are usually
large.}

\item{logo}{- character: path to a logo file. If NULL (default), we will take
package welovedatascience logo. If empty, no logo will be used. Only used
by revealjs format.}

\item{...}{- not yet used.}
}
\value{
inviisibly returns path to the generated presentation
}
\description{
This function generates a presentation (slides), currently targeting
powerpoint (other formats to come).
}
\examples{
\dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
slidfy(tar)
}
}
