% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tartree.r
\name{tartree}
\alias{tartree}
\title{tartree}
\usage{
tartree(
  data,
  tar_object = NULL,
  tarsum_object = NULL,
  target = NULL,
  decision_tree_maxdepth = 3,
  decision_tree_cp = 0,
  ...
)
}
\arguments{
\item{data}{data.frame or data.table}

\item{tar_object}{targeter object}

\item{tarsum_object}{targeter summary object}

\item{target}{character, target column name  - default: NULL and if tar_object is provided, target is taken from it}

\item{decision_tree_maxdepth}{integer, maximum depth of the tree - default: 3}

\item{decision_tree_cp}{numeric, complexity parameter - default: 0}

\item{...}{other parameters}
}
\value{
a targeter decision tree model on top of which a report can be generated with report function
}
\description{
Build a decision tree model following/based on a targeter analysis
}
\examples{
\dontrun{
data(adult)
tar_object <- targeter(adult, target = "ABOVE50K")
tarsum_object <- summary(tar_object)
tar_tree <- tartree(adult, tar_object = tar_object, tarsum_object = tarsum_object)
report(tar_tree)
}
}
